Update renamer.py
Added some safety and reliability changes (check if the "file" is actually a directory and skip it if so). There was never recursion, but it would break with folders.

+++ renamer.py
@@ -39,17 +39,20 @@ def blindrename(folder):
 
 	for file in os.listdir(folder):
 		old_name = path.join(folder, file)
-		if file == "keyfile.csv":
-			pass
-		elif file.startswith("."):
-			pass
-		else:
-			base = file.split(os.extsep, 1)
-			cloaked_name = "".join(random.choices(chars, k = 5))
-			new_name = path.join(folder, cloaked_name + "." + base[1])
-			print(old_name, " to ", new_name)
-			writer.writerow([old_name, new_name])
-			os.rename(old_name, new_name)
+		if path.isfile(old_name):
+			if file == "keyfile.csv":
+				pass
+			elif file.startswith("."):
+				pass
+			else:
+				base = file.split(os.extsep, 1)
+				# This ensures that there is a 1 index for files with no extension
+				base.append("")
+				cloaked_name = "".join(random.choices(chars, k = 5))
+				new_name = path.join(folder, cloaked_name + "." + base[1])
+				print(old_name, " to ", new_name)
+				writer.writerow([old_name, new_name])
+				os.rename(old_name, new_name)
 
 	print("Finished!")
